+++ src/Fixer/NullableParamStyleFixer.php
@@ -5,19 +5,27 @@ declare(strict_types = 1);
 namespace PhpCsFixerCustomFixers\Fixer;
 
 use PhpCsFixer\Fixer\ConfigurationDefinitionFixerInterface;
+use PhpCsFixer\Fixer\DeprecatedFixerInterface;
+use PhpCsFixer\Fixer\FunctionNotation\NullableTypeDeclarationForDefaultNullValueFixer;
 use PhpCsFixer\FixerConfiguration\FixerConfigurationResolver;
 use PhpCsFixer\FixerConfiguration\FixerOptionBuilder;
 use PhpCsFixer\FixerDefinition\CodeSample;
 use PhpCsFixer\FixerDefinition\FixerDefinition;
 use PhpCsFixer\FixerDefinition\FixerDefinitionInterface;
-use PhpCsFixer\Tokenizer\CT;
-use PhpCsFixer\Tokenizer\Token;
 use PhpCsFixer\Tokenizer\Tokens;
 
-final class NullableParamStyleFixer extends AbstractFixer implements ConfigurationDefinitionFixerInterface
+/**
+ * @deprecated use "nullable_type_declaration_for_default_null_value" instead
+ */
+final class NullableParamStyleFixer extends AbstractFixer implements ConfigurationDefinitionFixerInterface, DeprecatedFixerInterface
 {
-    /** @var string */
-    private $style = 'with_question_mark';
+    /** @var NullableTypeDeclarationForDefaultNullValueFixer */
+    private $fixer;
+
+    public function __construct()
+    {
+        $this->fixer = new NullableTypeDeclarationForDefaultNullValueFixer();
+    }
 
     public function getDefinition(): FixerDefinitionInterface
     {
@@ -34,7 +42,7 @@ function foo(int $x = null) {
     {
         return new FixerConfigurationResolver([
             (new FixerOptionBuilder('style', 'whether nullable parameter type should be prefixed or not with question mark'))
-                ->setDefault($this->style)
+                ->setDefault('with_question_mark')
                 ->setAllowedValues(['with_question_mark', 'without_question_mark'])
                 ->getOption(),
         ]);
@@ -42,76 +50,36 @@ function foo(int $x = null) {
 
     public function configure(?array $configuration = null): void
     {
-        $this->style = $configuration['style'] ?? $this->style;
+        if (isset($configuration['style']) && $configuration['style'] === 'without_question_mark') {
+            $this->fixer->configure(['use_nullable_type_declaration' => false]);
+        }
     }
 
     public function getPriority(): int
     {
-        // must be run before NoUnreachableDefaultArgumentValueFixer
-        return 1;
+        return $this->fixer->getPriority();
     }
 
     public function isCandidate(Tokens $tokens): bool
     {
-        return $tokens->isAllTokenKindsFound([T_FUNCTION]);
+        return $this->fixer->isCandidate($tokens);
     }
 
     public function isRisky(): bool
     {
-        return false;
+        return $this->fixer->isRisky();
     }
 
     public function fix(\SplFileInfo $file, Tokens $tokens): void
     {
-        for ($index = $tokens->count() - 1; $index > 0; $index--) {
-            if (!$tokens[$index]->isGivenKind(T_FUNCTION)) {
-                continue;
-            }
-
-            /** @var int $paramBlockStartIndex */
-            $paramBlockStartIndex = $tokens->getNextTokenOfKind($index, ['(']);
-
-            $paramBlockEndIndex = $tokens->findBlockEnd(Tokens::BLOCK_TYPE_PARENTHESIS_BRACE, $paramBlockStartIndex);
-
-            $this->fixFunction($tokens, $paramBlockStartIndex, $paramBlockEndIndex);
-        }
+        $this->fixer->fix($file, $tokens);
     }
 
-    private function fixFunction(Tokens $tokens, int $paramBlockStartIndex, int $paramBlockEndIndex): void
+    /**
+     * @return string[]
+     */
+    public function getSuccessorsNames(): array
     {
-        for ($i = $paramBlockEndIndex; $i > $paramBlockStartIndex; $i--) {
-            if (!$tokens[$i]->equals([T_STRING, 'null'], false)) {
-                continue;
-            }
-
-            /** @var int $variableIndex */
-            $variableIndex = $tokens->getPrevTokenOfKind($i, [[T_VARIABLE]]);
-
-            /** @var int $typeIndex */
-            $typeIndex = $tokens->getPrevMeaningfulToken($variableIndex);
-
-            if (!$tokens[$typeIndex]->isGivenKind([CT::T_ARRAY_TYPEHINT, T_CALLABLE, T_STRING])) {
-                continue;
-            }
-
-            /** @var int $separatorIndex */
-            $separatorIndex = $tokens->getPrevTokenOfKind($typeIndex, ['(', ',']);
-
-            /** @var int $nullableIndex */
-            $nullableIndex = $tokens->getNextMeaningfulToken($separatorIndex);
-
-            if ($this->style === 'with_question_mark') {
-                if (!$tokens[$nullableIndex]->isGivenKind(CT::T_NULLABLE_TYPE)) {
-                    $tokens->insertAt($nullableIndex, new Token([CT::T_NULLABLE_TYPE, '?']));
-                }
-                continue;
-            }
-
-            if (!$tokens[$nullableIndex]->isGivenKind(CT::T_NULLABLE_TYPE)) {
-                continue;
-            }
-
-            $tokens->clearAt($nullableIndex);
-        }
+        return [$this->fixer->getName()];
     }
 }